Adds a Makefile to fill dependencies for testing.
Depends upon pip being installed, but pip is pretty much
the standard nowadays and is just an easy_install away if
it isn't there.

The only dependency installed on to the system is
virtualenv which is used to make the other dependencies
local to the current environment.

Does not remove the need to install redis by hand, though
I am in favor of making that possible (using aptitude on
linux and brew on os x)

I look forward to cutting away at some of these
dependencies in further commits.