install_venv: don't use --no-site-packages with virtualenv
libvirt isn't listed in pip-requires because (a) it's not in PyPi and
(b) you always want to use libvirt's python bindings that matches the
version of libvirt installed on the system.

Currently, running nova-compute in virtualenv fails because libvirt
can't be imported. The --no-site-packages flag is what prevents this.
Everything seems to work fine without it. To verify, try deleting the
no-global-site-packages.txt from your .nova-venv.

Change-Id: I1df5e8e3c4426ca333c2d6b5b4fa8ece144dddf3