Remove python-novaclient dependency from nova.
Nova should not depend on python-novaclient. Remove it.

Note: In order to get unit tests to pass I had to add in
a temporary dependency on httplib2. This is due to the fact
that pip installing glance does not properly setup all
the dependencies. See LP Bug #975483 for details.

Fixes LP Bug #978992.

Change-Id: I50c00139f9dc303223601230fd962abac977cd5e