Fix yield bug for GlobalCheck
There was a bug where a GlobalCheck would report the same error four times
instead of once. This stemmed from a place in the code where we were iterating
a tuple, and on each iteration, yielding the whole tuple.

The result was that the 4-tuple, when returned, would generate four error
messages.

The solution is to yield the tuple once.

Change-Id: I87de3beded4b3c981d353a1647247ee84bbfccd0
Reviewed-on: https://review.openstack.org/30480
Reviewed-by: Monty Taylor <mordred@inaugust.com>
Approved: Sean Dague <sean@dague.net>
Reviewed-by: Sean Dague <sean@dague.net>
Tested-by: Jenkins