add check that python files have Apache 2.0 license stanza
I just had to -1 a proposed commit to Tempest which was missing
this, and found that 4 files in our tree didn't have it. We should
enforce that all the python files have this header when they are
in tree.

Skip files < 10 lines, which is needed for empty file support, and
to not bail on our Okay doctests.

Updated to include a check for an Apache looking LICENSE file as a
precondition.

Change-Id: Id3f9b6099301f998ce3185e9fd0ed6907e6e71f0