Disable H102 until ready to add new checks
We need to release a bug fix release, 0.5.5, which cannot contain any
new checks.  Currently there are several open bugs, and more testing
that is needed before we are ready to start adding new checks.

Since we are pinning hacking versions to <0.6, we can releases,
every time we add new checks we need to rev the minor release number.
But since we don't have extensive enough testing in this repository I
am not confident in our ability to add new working checks.

So until we are ready to cut 0.6 disable the check.

Change-Id: I84582fb41d859166395ba150c10e5a859a528950