Add check for end of title line of commit messages
The end of the title line of commit messages should not end with a
period. This is because space is limited to 50 chars and because we
treat this line like the subject line of emails.

Add this incredibly nitpicky style check to avoid any doubt over what
the correct style is.

Change-Id: I162ed090fbaef4f111792fdb89b31c2cab993b22