Add note about checks have community support
This project is only trying to enforce the agreed upon style guide, not
dictate it.  We do not want a new check added if there is any doubt the
check from the community.  Because this project has very a limited
number of reviewers we have to be extra careful about dictating style to the
rest of the community. Consensus in hacking reviewers is not community
consensus.

Change-Id: Id7aaefb8fb98b365f45c8eeb3aa53de02b9cb4e6