Strip leading/trailing whitespace from reference License
Some licenses are not indented at all, thereby missing a
leading space and failing the comparison. Given that
whitespace does not change the meaing of the license, it
is okay to remove it for the check.

Also adjust the print() statement which printed the
reference license as <license> and the actual license
as <apache2>, which is confusing for those who try
to understand what is going on.

Change-Id: I5d5c5fb67413848073d8ece2a723e98c111df1f1