Reword H404 description to match check
The check only checks that a multiline docstring starts
without a newline. The message "should start with a summary"
is misleading reviewers, as can be seen in e.g.

https://review.openstack.org/#/c/40114/

The "check for summary at beginning being less than 80 chars
and end with punctation" needs to be moved to a new hacking check
to preserve compatibility.

Change-Id: I694b0dc55706ccf8ada9bb132cc7efb3c49cd0d6