This resolves issues in the tests under PyPy
Adds ability to ignore hacking validations with noqa

Add allownoqa decorator that can be applied to any of the hacking
core functions.  This allows a using application to bypass that
particular hacking validation.  The reason this is being introduced
is because I need to bypass H231 and H233, both Python 3.x
validations.  I need to bypass these so we can run pep8 validations
on the nova XenServer plugins.  XenServer usese Python 2.4 in its
environment and therefore the code won't pass those two or future
Python 3.x validations.

It is understood that a patch to hacking won't be held up or reverted
if it breaks anything Python 2.4 related, including breaking of this
patch.

Change-Id: Iea3867b84212ab30bdd168b5dbd43c11a4dc535e
(cherry picked from commit 00348dfbe4a3b80c94415d7404a4be1f421bd46e)