Add a check for newline after docstring summary
The style guide says:

"""A multi line docstring has a one-line summary, less than 80 characters.

Then a new paragraph after a newline that explains in more detail any
general information about the function, class or method.
[...]

but hacking only checks that the text of the docstring starts
immediately after """.  It would be nice to check for the existence of
that empty line separating the summary from the rest of the comment.

Change-Id: I331a0f417f97784c083c06fb2fa18b384d8ed497
Closes-bug: #1249287