Add noqa support for H201 (bare except)
This'll be nice to have in high-level exception handlers that really do want
to catch *everything*

For example, Swift has some error handling code that is intended to
prevent exceptions from ever getting back to the user.  Because we
don't know what exception classes a third-party lib might raise
(including ones that might not inherit from Exception), it is not
possible to enumerate all of the possible exceptions so a bare
except: is the only reasonable option.

Change-Id: I4b44185df6b43316446b37225741523c7c829a3f