Disable logging correctly
logging.disable() function expects a log level as an int. The given
level is later compared against integer values (e. g.  INFO is 20). So,
when logging.disable() is given a string, it is later compared to int,
which is meaningless in python 2 and raises TypeError in python 3.

This change disables logging correctly, by passing maximal defined
log level (CRITICAL) to logging.disable().

Change-Id: I6d335fc127f04a5be65c51b33794cbb3db009aa0