Include rule numbers in HACKING.rst
Mention the hacking rule numbers in HACKING.rst. Make it easier for
readers to map the hacking rule numbers to HACKING.rst. This mapping
makes it easer to see what parts of HACKING.rst are not enforced by
hacking.

The notation of using a bulleted list and adding the rule number in square
brackets at the beginning of the line is taken from nova's list of local
HACKING rules.

Change-Id: Iddd2acbdd6c1f2409d16b55fe302383733f8575e