Clarify where the "following template" is
Reading through these docs, I was confused by the instruction to use
the "following template" which wasn't followed by a template.

Reading some more I realised that the template is a little further
down the page. I've made this more explicit by adding a link to the
template and changing the wording slightly.

Change-Id: I9f5086ae8102e42a7929e1467be834b20b64440a