remove H402
this really doesn't impact correctness of the code, or readability
of the comments. The rule really is that 1 line docstrings should
be a complete sentence. That's not machine checkable. Also, if
you are dealing with 80 col limits it's really hard to make these
complete sentences anyway if you only get 69 characters (79 - 4
space indent - 6 quote characters).

Change-Id: I9426644fa708e9d5563abe04dc4ad656dbeb3656