Add support for flake8 off_by_default for optional checks
This allows hacking checks to be added as optional which means
that they are assumed to be excluded and have to be explicitly
added as part of a 'select' entry in the [flake8]
stanza of tox.ini.

The goal is to allow checks to be staged into the projects instead
of requiring all of the projects to change to adhere to the
check prior to the check being added (which creates a window where
failing code could be added).  Instead checks can be staged in with
enforcement and once all (or the majority) of projects have enabled
it, the check can be changed to a required check.

One optional check is added as an example:
  H106: Don't put vim configuration in source files

Depends-On: I7f8ef22254703a53b73ccd048322191cfc677958
Change-Id: If188f85e8093974802bd2bacc2a72b2ae35ee5a1