Allow import redirections
When we move modules out of namespace packages we set up redirects to
import the new module under the old name. Python's import machinery
doesn't detect those things as modules, and the names don't show up in
sys.modules. However, if we look at what we actually got when we did the
import we can see that it is a module.

Change-Id: I4b42d081965b6d898b178cbe9232b47cfed17d8a