Add eventlet to test-requirements
eventlet isn't actually used by hacking, but its a common global import
(nova/__init__.py imports it) in OpenStack, so to make the hacking venvs
work with local hacks in repos that require eventlet, we just add it
here.

Since this is a test-requirement it shouldn't have any impact on
packaging of hacking.

Change-Id: I81098fb4f4debc2654d60ed28a02d64c5bbcedea
Closes-Bug: 1403270