Remove complex import rules
Per the discussion in [1], we are removing the import rules that
require introspection of the actual imported code due to the
complex and error-prone nature of those checks, and the fact that
some of them are simply impossible to implement sanely.

Checks that operate solely on the text of the import are left
alone because they should be less problematic.

[1]: http://lists.openstack.org/pipermail/openstack-dev/2014-December/052519.html

Change-Id: I43862a0d73706ee4b28951d5c8de6938c0cf5f27