tests: allow to run the same test for different code
If a function is registered to report different tests code (i.e. it
yields both H666 and H667), the test generator will fail as it does not
include the code number in the function name – therefore the function
will be generated twice with the same name. This patch fixes that.

Change-Id: I5394a066250f97882e895873164355e405b9c7cf