Add support for PEP257
This will install flake8_docstrings as a dependency when hacking is
installed using extras as hacking[pep257], thus adding pep257 as a
plugin to flake8.

Note that pep257 makes the entirety of the hacking.checks.docstrings
module redundant, as those checks are a subset of the checks from pep257
checks. For reference, here are hacking's docstring checks (H*) and
their PEP257 equivalents (D*):

    H401 = docstring should not start with a space
    H404: multi line docstring should start without a leading new line
    -> D210 = No whitespaces allowed surrounding docstring text

    H403: multi line docstrings should end on a new line
    -> D209: Multi-line docstring closing quotes should be on a separate line

    H405: multi line docstring summary not separated with an empty line
    -> D205: 1 blank line required between summary line and description

Change-Id: I80761792b7c9b947a775e68050c4ba2758d9a72e
Related-Bug: 1501544
Depends-On: I60adf0dca4aa32f4ef6bca61250b375c8a3703c6