Add optional H203 to check that assertIs(Not)None is used
...instead of assertEqual(None, ...) or assertIs(None, ...) and their
negations.

Apparently that's a thing that people want? At any rate, 30+ projects
have this bug marked "Fix Released". However, without a check to keep
them from regressing, there is a long tail of follow-up patches that
distracts developers and reviewers from addressing user-facing bugs. See
the following Nova changes, spread out over the course of a year, for an
example:

    If3824356ddf4e6e2d91f6bc2fbfa41946d8463cc
    I09f38e219931e0d7ad27f04861d1ebbc3b5e2c5f
    Iee1379b941f93388900e89388676000b845fc8fc
    I0d38a82e78fbe94657ab9a71c08422007843d179
    I9316c0b125aa87b6ebfa996a559c3551093ea711
    I406ea23b0e78f45f16306813e4111a95716cd6b0
    Ic8cb1192e001409d827c8da55fe536681895944b

Add a check for it in the interest of not wasting reviewer time or
having more code churn than necessary.

Change-Id: Iad65cb6399f4f933cbd9f503c88ce144387d39b5
Related-Bug: #1280522