Remove the PBR cap
Unlike most of the other OpenStack projects hacking is not automatically
syncronsed to projecst via the proposal-bot.  My understanding of this
is because new hacking tests could cause unexpected gate failures which
it was automatically syncronised and/or did not always take into account
project style.

The downside of this is that we have alarge number of projects still
using older hacking releases.

This has come to a head with the release of pbr 2.0.0 which conflicts
with 0.9.x and 0.10.x hacking branches.

While it isn't possible to get all projecst up to the newest hacking it
it is possible to remove the cap on PBR and allow project that still
want older hacking releasee to work with the new PBR release.

Change-Id: Iabf27cc0648c12c3c090f01facd15c3ec52a4861
Related-Bug: 1668848