Release notes for 0.14.0
There has not been a hacking release in some time. This release is
done at the very beginning of the Queens cycle to reduce disruption as
much as possible and give plenty of time to solve any issue that may
arise or enabling new rules that are disabled by default.

Change-Id: If9a54a0b2c07f91ff2f503f1159ff0d23946e2d9