Release notes for 1.1.0
As we have 2 new changes and request to release
the hacking, this commit adds the release notes
for next release.

Usually we do not add the release note for hacking
changes but it is good to have release notes per
release to reflect the glance of changes happening in
release.

Change-Id: I607ea48e9539e9bb2e2115dec503ed7283fb84af