Fix 'ref' format errors in README file
While releasing new version in pypi, it perform the
validation on the package and give error if there is any
formatting issue.

README file is being used to upload on pypi and so validation
check fails[1] on current README due to 'ref' used for internal
doc link. Those links would not be available for outside the built doc.
This can be checked locally by:
    python setup.py check -s -r

This commit fixes those format issue by removing the ref and provide the
external links.

[1] http://logs.openstack.org/17/622317/1/check/test-release-openstack/a24116d/job-output.txt.gz#_2018-12-05_06_16_03_329613

Change-Id: I70c995b715184c2cc70b494c185c83b8190629db