Bump flake8 version to something modern
The flake8 version we're using is a good 2 years old and is in turn
using an old version of pycodestyle. This is causing the following
warnings:

  FutureWarning: Possible nested set at position 1
    EXTRANEOUS_WHITESPACE_REGEX = re.compile(r'[[({] | []}),;:]')

Bump to the latest and greatest so we get these fixes. This involves
dropping support for local-checks, which don't seem to work in this
version. Fortunately, in the intervening time, flake8 has grown the
ability to do local checks all by its lonesome. The tests also need to
be reworked and unfortunately are much slower now. This is because
flake8 3.x's API is file-based [1] and without rooting around the guts
of flake8, it's not practical to do things any other way. Hopefully
flake8 4.x won't have such issues.

[1] https://gitlab.com/pycqa/flake8/issues/545

Change-Id: I695ff02a6970663add10caf7f16a66abf9d1239d
Signed-off-by: Stephen Finucane <sfinucan@redhat.com>
Sem-Ver: api-break