Disable old-style-class check by default
In Python3, all objects inherit 'object' and there is no need to
inherit 'object' in class declarations as we do in python2.
pylint 2.0 added a new check useless-object-inheritance and
if this check is enabled in pylint, we need to disable the
old-style-class check in hacking. We are now in py3 era,
so it sounds good to disable it by default.

Change-Id: Iaebcf267de9ef826b8cc387d1fa1f2eccff2c050