migrate testing to ubuntu focal
As per victoria cycle testing runtime and community goal[1]
we need to migrate upstream CI/CD to Ubuntu Focal(20.04).

Fixing:
- bug#1886296
Bump the flake8 to 3.8.0 as min so that pyflakes 2.1.1 as min version taken
care to run pep8 jobs on py3.8 which is default python vesion in ubuntu focal.
- https://github.com/PyCQA/flake8/blob/3.8.0/setup.cfg#L44

Story: #2007865
Task: #40212

Closes-Bug: #1886296
[1] https://governance.openstack.org/tc/goals/selected/victoria/migrate-ci-cd-jobs-to-ubuntu-focal.html

Change-Id: I19b13d5d63bce0f6b646972b09cfe8e8d7c22adc