Handle formatting of args in string interpolation check
We have added a check to make sure format strings and any args are being
passed in to be formatted as needed by the logger. The check did not
take into account string formatting that may be needed in preparing the
value args. There are some cases where this is needed, and we should not
be raising an error if the string formatting is being done there.

This adds a new test string and updates the hacking check to account for
these cases so the new hacking release does not raise unnecessary
errors.

Change-Id: Ie4e3f95d24dc954deb7336835a6f9b7bf54c9674
Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>