Bump flake8 to 5.0 to pick up fixes
There is a bug [1] in older flake8 that we need to pick up. In
particular, Nova has at least one line longer than 80 characters in it
[2] because of that bug. Our own unit tests start breaking with the
new flake8, so we need to fix whitespace for one `if`, and start
calling flake8 with `--enable-extentions` to force flake8 to load
extensions decorated with @off_by_default.

[1] https://github.com/PyCQA/flake8/issues/1534
[2] https://opendev.org/openstack/nova/src/commit/5c32d5efe1e1aece48b680474617113c61a248d5/nova/image/glance.py#L392

Change-Id: I7799049924b5a2cde867c88c03154c36831c9670