Remove Python 2->3 checks
These were not running on Python 3 codebases which is everything now.

Note that 'assertEquals'/'assertNotEquals', which is checked for by one
test, is still present in some codebases, despite it being removed in
Python 3. This is because this variant is added by testtools, which
adds its own deprecation warning that we don't need to duplicate.

Change-Id: I3f15376db78886b7eecad3fdfdd363c6ea5d19c6
Signed-off-by: Stephen Finucane <stephenfin@redhat.com>